@@ -4,7 +4,7 @@ import { ServiceInstance } from '../../../../shared/datamodel/k8s/service-instan
 import { ServiceInstancesService } from '../../../../service-instances/service-instances.service';
 import { ServiceBindingsService } from '../../../../service-bindings/service-bindings.service';
 import { ServiceBinding } from '../../../../shared/datamodel/k8s/service-binding';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { InstanceBindingInfo } from '../../../../shared/datamodel/instance-binding-info';
 import * as luigiClient from '@kyma-project/luigi-client';
 import { ServiceBindingList } from '../../../../shared/datamodel/k8s/service-binding-list';
