@@ -1,35 +1,57 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import {
-  ServiceBindingUsage, IServiceBindingUsage, IServiceBindingUsageSpec,
-  ILocalReferenceByName, ILocalReferenceByKindAndName, IServiceBindingUsageList
+  ServiceBindingUsage,
+  IServiceBindingUsage,
+  IServiceBindingUsageSpec,
+  ILocalReferenceByName,
+  ILocalReferenceByKindAndName,
+  IServiceBindingUsageList,
 } from '../shared/datamodel/k8s/service-binding-usage';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { AppConfig } from '../app.config';
 import { IMetaData } from '../shared/datamodel/k8s/generic/meta-data';
 
-
-
 @Injectable()
 export class ServiceBindingUsagesService {
+  constructor(private http: HttpClient) {}
 
-  constructor(private http: HttpClient) { }
-
-  public createServiceBindingUsage(serviceBindingUsage: ServiceBindingUsage, token: string): Observable<ServiceBindingUsage> {
+  public createServiceBindingUsage(
+    serviceBindingUsage: ServiceBindingUsage,
+    token: string,
+  ): Observable<ServiceBindingUsage> {
     const httpOptions = this.getHTTPOptions(token);
-    const url = `${AppConfig.serviceBindingUsageUrl}/namespaces/${serviceBindingUsage.metadata.namespace}/servicebindingusages`;
-    return this.http.post<ServiceBindingUsage>(url, serviceBindingUsage, httpOptions);
+    const url = `${AppConfig.serviceBindingUsageUrl}/namespaces/${
+      serviceBindingUsage.metadata.namespace
+    }/servicebindingusages`;
+    return this.http.post<ServiceBindingUsage>(
+      url,
+      serviceBindingUsage,
+      httpOptions,
+    );
   }
 
-  public deleteServiceBindingUsage(name: string, namespace: string, token: string): Observable<ServiceBindingUsage> {
+  public deleteServiceBindingUsage(
+    name: string,
+    namespace: string,
+    token: string,
+  ): Observable<ServiceBindingUsage> {
     const httpOptions = this.getHTTPOptions(token);
-    const url = `${AppConfig.serviceBindingUsageUrl}/namespaces/${namespace}/servicebindingusages/${name}`;
+    const url = `${
+      AppConfig.serviceBindingUsageUrl
+    }/namespaces/${namespace}/servicebindingusages/${name}`;
     return this.http.delete<ServiceBindingUsage>(url, httpOptions);
   }
 
-  public getServiceBindingUsages(namespace: string, token: string, params: object = {}): Observable<IServiceBindingUsageList> {
+  public getServiceBindingUsages(
+    namespace: string,
+    token: string,
+    params: object = {},
+  ): Observable<IServiceBindingUsageList> {
     const httpOptions = this.getHTTPOptions(token, params);
-    const url = `${AppConfig.serviceBindingUsageUrl}/namespaces/${namespace}/servicebindingusages`;
+    const url = `${
+      AppConfig.serviceBindingUsageUrl
+    }/namespaces/${namespace}/servicebindingusages`;
     return this.http.get<IServiceBindingUsageList>(url, httpOptions);
   }
 
@@ -38,9 +60,9 @@ export class ServiceBindingUsagesService {
     httpHeaders = {
       headers: new HttpHeaders({
         'Content-Type': 'application/json',
-        'Authorization': `Bearer ${token}`
+        Authorization: `Bearer ${token}`,
       }),
-      params: parameters
+      params: parameters,
     };
     return httpHeaders;
   }
@@ -48,19 +70,19 @@ export class ServiceBindingUsagesService {
   initializeServiceBindingUsage(): ServiceBindingUsage {
     const md: IMetaData = {
       name: '',
-      namespace: ''
+      namespace: '',
     };
     const sbName: ILocalReferenceByName = {
       name: '',
-    }
+    };
     const ub: ILocalReferenceByKindAndName = {
       name: '',
       kind: '',
-    }
+    };
     const sp: IServiceBindingUsageSpec = {
       serviceBindingRef: sbName,
-      usedBy: ub 
-    }
+      usedBy: ub,
+    };
     const serviceBindingUsage = new ServiceBindingUsage({
       kind: 'ServiceBindingUsage',
       apiVersion: 'servicecatalog.kyma.cx/v1alpha1',
